Add YHSearchBar, remove YHTextField
